backend/requirements.txt
# Dependencias del backend
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
bcrypt
python-dotenv
python-multipart
httpx
pytest
pytest-asyncio
